The working version of the alternatives to get.
