+++ lib/get/onValueType.js
@@ -7,6 +7,7 @@ var $error = require("./../types/error");
 var onError = require("./onError");
 var onValue = require("./onValue");
 var onMissing = require("./onMissing");
+var isMaterialized = require("./util/isMaterialzed");
 
 /**
  * When we land on a valueType (or nothing) then we need to report it out to
