+++ lib/get/walkJSON.js
@@ -1,7 +1,6 @@
 var followReference = require("./followReference");
 var onValue = require("./onValue");
 var onValueType = require("./onValueType");
-var isMaterialized = require("./util/isMaterialzed");
 var isExpired = require("./util/isExpired");
 var iterateKeySet = require("falcor-path-utils").iterateKeySet;
 var $ref = require("./../types/ref");
