+++ lib/get/walkPath.js
@@ -3,11 +3,6 @@ var onError = require("./onError");
 var onMissing = require("./onMissing");
 var onValue = require("./onValue");
 var onValueType = require("./onValueType");
-var lru = require("./util/lru");
-var hardLink = require("./util/hardlink");
-var isMaterialized = require("./util/isMaterialzed");
-var removeHardlink = hardLink.remove;
-var splice = lru.splice;
 var isExpired = require("./util/isExpired");
 var iterateKeySet = require("falcor-path-utils").iterateKeySet;
 var $ref = require("./../types/ref");
