+++ test/get-core/index.js
@@ -1,3 +1,3 @@
 describe('Get', function() {
-    require('./get.core.spec');
+    require('./get.core.pathMap.spec');
 });
